Cleaning up, consolidating properties to single file
