+++ src/main/java/com/gtnexus/appxpress/pmdocgen/adapter/PrimitiveDisplayFunction.java
@@ -1,4 +1,4 @@
-package com.gtnexus.appxpress.pmdocgen;
+package com.gtnexus.appxpress.pmdocgen.adapter;
 import com.google.common.base.Function;
 
 public abstract class PrimitiveDisplayFunction<F> implements Function<F, String> {
